requirements.txt
numpy==1.24.3
natsort
matplotlib
scikit-learn
scikit-image
yacs
joblib
h5py
tqdm
tensorboard
einops
gdown
addict
future
lmdb
pyyaml
requests
scipy
yapf
lpips
pytorch_msssim
nvitop
opencv-python
thop